chore(FileList): add new components; factor out common util funcs
